Make the existing alerter tests into unit tests
The existing tests are integration-ish. This adds tests based on the existing
tests which test the alerter in a more functional way—setting it up, giving it
input, and checking its output.

These new tests are more confined in what they test and generally simpler,
while the existing tests are more holistic and able to perform more checks (for
example, they can verify that behavior is correct with regard to RMCs). Both
sets of tests have value.