alerter tests: Rename own_node to this
`own_node` was used in an attempt to reduce confusion (in contrast with
`alerter`) when a test involved different nodes sending alerts. But, as Damian
Straszak pointed out, _node_ usually has a more holistic meaning. `this` also
seems appropriate, since the alerter is the relevant component being tested in
these contexts.