Send requests for missing data to random peers
Change how both CoordRequests and ParentRequests are managed:

CoordRequests:

1. CoordRequests are sent to a configurable number of random nodes
   (default 3, 3, 3, 1, 1, 1, 1, 1, ...).
2. Make the initial retries more aggressive with a default schedule of
   100, 1000, 3000, 6000, 9000, ... ms.

These changes aim at making the initial response faster in case the
network is spotty but not completely down.

ParentRequests:

ParentRequests are sent to a single random node from the first try
instead of the creator. Given that this mostly happens when we suspect
someone of cheating it might not even make sense to query the creator.